requirements.txt
discord.py==2.5.2
dotenv==0.9.9
asyncio==3.4.3
mysql-connector-python==8.3.0
